requirements.txt
boto==2.20.1
python-gnupg==0.3.6
PyYAML==3.10
